refactor(template-advanced): split main.py into focused modules, use EmailStr
The single 505-line main.py mixed nine concerns; it now teaches the
canonical layout:
- config.py    Settings loaded from the environment
- database.py  engine, session factory, get_db dependency
- models.py    SQLAlchemy ORM models
- schemas.py   Pydantic request/response models
- auth.py      JWT tokens, password hashing, current-user dependency
- realtime.py  WebSocket connection manager
- main.py      app, middleware, rate limiter, endpoints (305 lines)

Pure refactor verified green before and after: 31 tests in-place and
31 in a freshly generated (renamed) project, whose imports the CLI
rewriter handles across all seven files. ruff src = ["src"] keeps
first-party import grouping correct under any project name.

Also replaces the homemade email regex with pydantic EmailStr
(email-validator==2.3.0): was RED - "user name@example.com" with a
space registered successfully; now 422.
