+++ packages/template-advanced/pyproject.toml
@@ -17,6 +17,7 @@ dependencies = [
   "bcrypt==5.0.0",
   "slowapi==0.1.10",
   "python-multipart==0.0.32",
+  "email-validator==2.3.0",
 ]
 
 [tool.hatch.build.targets.wheel]
@@ -32,6 +33,7 @@ dev = [
 [tool.ruff]
 target-version = "py311"
 line-length = 120
+src = ["src"]
 
 [tool.ruff.lint]
 select = ["E", "F", "I"]
