+++ packages/template-advanced/src/advanced/auth.py
@@ -1,0 +1,78 @@
+import datetime
+from typing import Annotated
+
+import bcrypt
+import jwt
+from fastapi import Depends, HTTPException, status
+from fastapi.security import OAuth2PasswordBearer
+from sqlalchemy import select
+from sqlalchemy.orm import Session
+
+from advanced.config import settings
+from advanced.database import get_db
+from advanced.models import User
+
+# The JWT signing key comes from the environment via Settings (see .env_dev)
+ALGORITHM = "HS256"
+ACCESS_TOKEN_EXPIRE_MINUTES = 30
+
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/auth/login")
+
+
+def verify_password(plain_password: str, hashed_password: str) -> bool:
+    # bcrypt has a 72-byte limit, truncate if necessary
+    password_bytes = plain_password.encode("utf-8")[:72]
+    return bcrypt.checkpw(password_bytes, hashed_password.encode("utf-8"))
+
+
+def get_password_hash(password: str) -> str:
+    # bcrypt has a 72-byte limit, truncate if necessary
+    password_bytes = password.encode("utf-8")[:72]
+    return bcrypt.hashpw(password_bytes, bcrypt.gensalt()).decode("utf-8")
+
+
+def create_access_token(data: dict, expires_delta: datetime.timedelta):
+    """
+    Create JWT access token
+    TODO: Add refresh token support:
+    - Separate refresh token with longer expiry
+    - Token revocation list (blacklist)
+    - Different permissions in token payload
+    """
+    to_encode = data.copy()
+    to_encode.update({"exp": datetime.datetime.now(datetime.UTC) + expires_delta})
+    return jwt.encode(to_encode, settings.secret_key, algorithm=ALGORITHM)
+
+
+def get_user_by_username(db: Session, username: str) -> User | None:
+    return db.execute(select(User).where(User.username == username)).scalars().first()
+
+
+async def get_current_user(
+    token: Annotated[str, Depends(oauth2_scheme)],
+    db: Annotated[Session, Depends(get_db)],
+) -> User:
+    """
+    Get current authenticated user
+    TODO: Add role-based access control:
+    - Check user permissions
+    - Support for API keys
+    - Session management
+    """
+    credentials_exception = HTTPException(
+        status_code=status.HTTP_401_UNAUTHORIZED,
+        detail="Could not validate credentials",
+        headers={"WWW-Authenticate": "Bearer"},
+    )
+    try:
+        payload = jwt.decode(token, settings.secret_key, algorithms=[ALGORITHM])
+        username: str = payload.get("sub")
+        if username is None:
+            raise credentials_exception
+    except jwt.InvalidTokenError:
+        raise credentials_exception
+
+    user = get_user_by_username(db, username)
+    if user is None:
+        raise credentials_exception
+    return user
