+++ packages/template-advanced/src/advanced/schemas.py
@@ -1,0 +1,46 @@
+import datetime
+
+from pydantic import BaseModel, ConfigDict, EmailStr, Field, field_validator
+
+
+class UserCreate(BaseModel):
+    username: str = Field(..., min_length=3, max_length=50)
+    email: EmailStr
+    password: str = Field(..., min_length=6)
+
+
+class UserResponse(BaseModel):
+    model_config = ConfigDict(from_attributes=True)
+
+    id: int
+    username: str
+    email: str
+    is_active: bool
+    created_at: datetime.datetime
+
+
+class ProductCreate(BaseModel):
+    name: str = Field(..., min_length=1, max_length=100)
+    description: str | None = Field(None, max_length=500)
+    price: float = Field(..., gt=0)
+
+
+class ProductResponse(BaseModel):
+    model_config = ConfigDict(from_attributes=True)
+
+    id: int
+    name: str
+    description: str | None
+    price: float
+    created_at: datetime.datetime
+
+    @field_validator("price", mode="before")
+    @classmethod
+    def cents_to_dollars(cls, value: float) -> float:
+        # Product.price is stored as integer cents
+        return value / 100
+
+
+class Token(BaseModel):
+    access_token: str
+    token_type: str
